requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
discord.py==2.2.3
Django==4.2.1
Faker==18.9.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
webpreview==1.7.2
yarl==1.9.2
